Added validations & integrated swagger

+++ src/services/sites.service.ts
@@ -11,6 +11,8 @@ export class SitesService {
   ) {}
 
   async create(url: string, label: string) {
+    const existingSite = await this.findByURL(url);
+    if (existingSite) return { message: 'This site already exists' };
     const site = this.sitesRepository.create({
       url,
       label,
@@ -49,7 +51,18 @@ export class SitesService {
   }
 
   async deleteById(id: string) {
+    const site = await this.findById(id);
+    if (!site) return { message: 'Site not found' };
     await this.sitesRepository.delete(id);
     return { message: 'Site successfully deleted' };
+  }
+
+  async deleteAll() {
+    const sites = await this.findAll();
+    if (sites && sites.length > 0) {
+      await this.sitesRepository.remove(sites);
+      return { message: 'Sites removed' };
+    }
+    return { message: "There's no site in DB" };
   }
 }
